Remove dtype from generics (#590)
The introduction of `DataType` as generic in `NDArray` caused a lot of confusion and unreadable code. To vastly simplify the codebase and the compiler errors, we are removing `dtype` as generic. We expect this to have negligible impact on usability since most arrays are of type `float32`.

`NDArray<DataType, Rank>` becomes  `NDArray<Rank>`.

This change is breaking typescript users that use generics. JavaScript users should not be affected.

To further improve compiler error messages, we are replacing `Rank` from a union string to an enum. Before, in a compiler error message, or in a code editor, an array of unknown rank showed up as `NDArray<'0'|'1'|'2'|'3'|'4'|'higher'>`. Now it is `NDArray<Rank>`. `NDArray<'3'>` now shows up as `NDArray<Rank.R3>`.

Internal changes
 - Removed all instance and static methods of subclasses of NDArray (e.g. Scalar, Array1D etc). This improves the compiler inference and error messages.