Introduce src/contrib/data (#639)
This subpackage provides Dataset, DataStream, and related subclasses and utilities.

It provides for streaming data loading and preprocessing, with the intent of feeding machine learning models with data for both training and evaluation.

This cannot yet replace src/data, nor does it yet support the needs of the current deeplearn.js demos. It's just an initial commit, moving the current version of the code unmodified from its private repo. Deeper integration with deeplearn.js will follow.