User lower camel case for initializer / regularizer / constraints at the API surface. (#52)
At the API call site, we use lower camel case for the `InitializerIdentifier` and accepted string values. We also accept the upper case version.

For example: `'Ones'` => `'ones'`.

This does not change how serialization works, or any of the maps. At the call site of `getInitializer` in initializers.ts, we map from the lower camel case to the upper camel case version to be used in the `ConfigDict` internal representation.

This comes from the discussion offline about the difficulty of the "proper" approach, which is to actually change the registry. The problem with that comes from assumptions that are baked in
1) Constructor.name maps back the registry key. This is not true if I switch to lower camel everywhere.
2) The constructor.name maps all the way back to the keras python symbol ([code here](https://github.com/tensorflow/tfjs-layers/blob/master/src/utils/serialization_utils.ts#L128)), which is no longer true. This means we'd have to remove this special case line and map back.

Given that this is quite risky and that we want to redo serialization, the patch fix I've sent in this PR seems like the least risky.

This PR also:
- Throws an error when not defining a recurrent initializer before we fail saying "orthogonal doesnt exist"