Simplify the calling contract of ClassNameMap.register (#146)
Simplify the calling convention for ClassNameMap.register, by enforcing both class-type and instance-type contracts on Serializable.

The current requirements for Serializables.
Define a static member className that returns the json compatible format for python/js interop.
Override the static member fromConfig if special handling is required
Define a non-static getClassName that returns the classobject.className field.  (This is the remaining annoyance of the approach, but we can't find an appropriate workaround.)
Define a non-static getConfig.