Add a Keyword Argument type. (#163)
Adds a Kwargs type to represent keyword argument bundles, typically passed around call() methods of Layers.  This is a replacing an "any" type, and is itself defined with an any, so its not significantly tighter.  It does enforce a string-indexed to any mapping.  Overtime we might be able to further tighten this, and at least will have a single point of edit.