Remove Enums in favor of TS types with guard functions (#164)
We have already changed most of the Enums to type unions.  However we still had some residue of the old Enum approach, mainly affecting the serialization needs.  This removes that cruft, simplifying the code several places.  Additionally, it refactors the helper guard functions to a single place and standardizes on only listing the allowed values (not undefined nor null), while still handling those properly in checking and error messages.