Apply tidy systematically across tfjs-layers. (#186)
All of layers code should be tidy'd now.

We don't use @operation, since it only works on static class methods and we're entirely in either free functions or non-static instance members.

We currently overly tidy -- ie even functions that create no intermediate tensors are tidied.  Our policy going forward, to remove a tidy, you must add a test that proves there's no leakage.  That way if the implementation of a non-tidy'd functiob changes, we have a safety net to catch it require the reintroduction of the tidy.