Make sure tf.fromPixels() gets called after the DOM is ready (#1064)
BUG Since a newer version of Chrome, you have to create the canvas element when the DOM is ready in order to get a 2d rendering context. This PR postpones the creation of a fromPixelsCanvas element until the user calls fromPixels(). This makes the webcam-transfer-learning example work.

Fixes tensorflow/tfjs#313