Generate an ES Module bundle, too. (#368)
DEV Fixes #367. This produces a single ES module bundle with no dependencies. If you prefer, it might be slightly better to retain the dependencies on the other Tensorflow packages (*e.g.*, tfjs-converter). That way if you import those other packages directly as well as tfjs, you don’t get duplicate code. (This would require that the other Tensorflow packages likewise publish an ES module bundle as in this PR.)