Update benchmarks to use latest published layers (#235)
Fixes https://github.com/tensorflow/tfjs/issues/416

Benchmarks in layer were out of date and not working.

Part of the solution is to not `yarn link` to the current master branch because it creates duplicate backend and hard to keep in sync. Instead we link to specific version on npm. For those interested in making changes to layers and re-running benchmarks can use `yalc/yarn` locally on their machine.

Also show tf.js core and layers version numbers on the benchmarks page since there is confusion as to which version the benchmarks are using. See version field in https://github.com/tensorflow/tfjs/issues/416 which points to TF Python version, not the tfjs version.

BUG