Improve precision for Safari by using float16 on iOS and encoding floats as bytes before downloading. (#978)
BUG

This PR enables iOS Safari and Desktop Safari to have higher floating point precision (vs now which is fixed point precision). This means we also support numbers larger than 20,000 (and less than -20,000). To do this, we must use different types of textures for uploading, rendering, and downloading.

On all platforms, uploading to float32 textures is supported.

|         | Render           | Download  |
| ------------- |:-------------:| -----:|
| Chrome / Firefox      | float32 | float32 |
| Safari Desktop      | float32      |   bytes |
| iOS Safari | float16      |    bytes |

To do this, we remove the `WEBGL_FLOAT_TEXTURE_ENABLED` bit and introduce two new bits, `WEBGL_RENDER_FLOAT32_ENABLED` and `WEBGL_DOWNLOAD_FLOAT_ENABLED` which enable each of the respective code paths depending on the device.

Because we have multiple types of textures, the texture manager now knows about the type of textures, and will cache based on the texture type. I introduced a separation of `LogicalTextureType` which includes "upload", "render" and "download", so from the perspective of the WebGL backend we ask the texture manager for a texture of this type. The texture manager will convert that to a `PhysicalTextureType` depending on environment flags, and create a texture based on that type. This means we maximize cache utilization across flag settings.

Because we can't download floating point textures directly, we have to first pass it through an IEEE 32 bit float encoder program which samples from a floating point texture and writes to an unsigned byte texture. In JavaScript, we simply wrap that `ArrayBuffer` in a `Float32Array` to decode on the CPU-side. This only happens when the user asks to download a Tensor from the GPU (via readSync in BackendWebGL).

This PR also:
- Turns off Firefox testing locally. This speeds up testing for local development.
- On Browserstack, we now run Chrome, Firefox, Safari, and iOS Safari.
- For iOS devices (render float16) we reduce the precision in unit tests to 1e-1. Late, we can see if we can improve precision.
- Turns off a few tests for non-Chrome browsers. I filed issues to track turning them back on (all IO handler related).
- Refactors gpgpu_context so that the texture configuration is computed a single time when the context is created, instead of computing them whenever needed. This cleans up the file a lot.
- Removes all the code paths for packing.
- Swaps `isNaN` for a new implementation that works on all platforms (tested on windows edge).
- Reduces the size of some values in unit tests to pass in 1e-1 setting on iOS.