Add 6D tensors. (#1084)
FEATURE
As described in the comments on this issue tensorflow/tfjs#161 I think having a 6D sampler and 6D GPU transpose will be the best way to fully support the batchToSpaceND op. This PR adds a 6D tensor type, GPU support for 6D tensors, GPU support for 6D transpose and tests for a few various ops.