Fixing various nan related issues impacted by floating point precision and NaN (#1116)
BUG

Fixes https://github.com/tensorflow/tfjs/issues/291

On the newer macs we had issues computing pow, mod, acosh, and atanh due to various issues related to NaN not being properly handled at the GPU level. 

In the case of **acosh** and **mod** the problem stemmed from Div not successfully dividing 3/3 (or 6/6, 7/7, etc). In those cases the result would be .99999 and when floor was applied we got 0 instead of 1. 

In the case of **pow** the problem stemmed from **log** not catching NaN for negative numbers.

Lastly, **atanh** failed due to not checking the inputs for values outside of the range [-1, 1]. For those values **atanh** needed to just return NaN directly.