Remove dependency on tfjs-core/dist. (#166)
We should never import from `dist/`. If we need something in `dist/` it should be exported as part of the public API.

I opted for duplicating types that will never change instead of exporting from core and importing. It's better to have fewer symbols and decouple the two libraries for simple types.