Add gradients to `oneHot`,  `argMax`, `argMin`, `greaterEqual` (#1218)
FEATURE

* Add gradient implementation for `oneHot`

`oneHot` op is not differentiable and thus should have zero gradient
with the same shape and dtype as its input.

* Add gradient implementation for `argMax`

`argMax` op is not differentiable and thus should have zero gradient
with the same shape and dtype as its input.

* Copy `begin` argument of `slice` to prevent change

`tf.split()`, `tf.unstack()` modify `begin` argument in a loop and thus
change its value for the backpropagation. Make a copy of `begin` in
`slice` to prevent such changes.

* Add gradient implementation for `argMin`

`argMin` op is not differentiable and thus should have zero gradient
with the same shape and dtype as its input.

* Add gradient implementation for `greaterEqual`

`greaterEqual` op is not differentiable and thus should have zero
gradient with the same shape and dtype as its input.