Deep zip (#52)
Provides a function to zip nested structures together in a recursive manner.

To date, batching and computation of statistics have been supported only on purely tabular data, i.e. where each element is a flat array or dict. However, we must be able to do these things when data elements are nested structures (#22, #32). The utility provided in this PR will make both of those changes much simpler in future PRs.

Note the structure of the code is very similar to the existing deepMap().