Column-major batching of nested structures. (#55)
Removes special types to represent batches, and implements columnar batching of nested objects.

Remove the BatchDataset class. The result of batching is now a regular Dataset, not a BatchDataset, and it's up to the user to keep track of the batch nature (i.e. typically that dimension 0 is the batch dimension).

Remove BatchArray and DatasetBatch types.

Move ElementArray and TabularRecord types to statistics.ts, which is the only remaining usage (also to be expunged per #32).

Clarify distinction between row-major batches and column-major batches.

Reimplement Dataset.batch() to handle elements with nested structure. First performs row-wise batching, and then applies deepZip() to each batch with a custom zip function to combine the leaves into batches (represented as Tensors when possible).

Aside: removes IteratorProperties per offline discussion.

Fixes #22. Fixes #23.