Allow squeeze to use negative index (#1430)
Issue in focus: tensorflow/tfjs#940

#### Description

`squeeze()` didn't support the use of negative index which is supported in tensorflow for Python. Added tests for the same in `util_test` and `tensor_test`.
The previous implementation didn't allow for random arrangement of values of axis. For example if axes is set to `[2, 3]`, it would remove them both but if you pass it as `[3, 2]` it would only remove 3 and ignore 2 entirely. I tried the same with tensorflow for python in this [Google Colab Notebook](https://colab.research.google.com/drive/1jzyuRFXfIVrK50jI7vDrVuW5dYpiYw0V) and this too was inconsistent with tfjs as tensorflow python allows for removal of axis irrespective of order of input if it's allowed. This inconsistency was removed when adding the support for negative index ( it was totally unintentional but resulted from my implementation to allow negative index ).

Clarification required for following topics:
* I am not sure whether the test messages are satisfactory for the new tests added - Please let me know if this can be improved and I'll implement the same as soon as possible.
* The random input order support was a result of my judgement where i thought `[-1, -2]` was intuitively much better for a programmer but if we just replaced it with the positive index of same, it would have been `[5, 4]` which due to it's lack of ascending order would have ignored 4 completely. If this needs to changed, please let me know and I'll come up with another implementation to change this behavior.

Tests Ran:
* [x] yarn test
* [x] yarn lint

BUG