Fix condition for expensive reshape with even columns (#1441)
BUG

Packed reshape when both, source and destination shape, have even columns is not free - with current 2x2 packing, pixel's .ba values are the next (logical) row of .rg values. If row length changes,
different values are expected in .ba values.

Added test that exposes the issue and the fix.