Enable eager mode in TF when benchmarking (#396)
Change the python benchmarking code to use eager mode and `tf.keras` for better comparison with tf.js. This makes a significant difference in the `predict()` and `fit()` time for recurrent cells such as LSTMCell (see attached benchmark at the end). Also set PROD flag to true when running the benchmark in the browser.

**Other stuff**
- Evaluate the EPSILON flag lazily as opposed to at module-import time. This allows users to control PAGING and other flags by setting the PROD flag.
- Disable karma instrumentation by default (aligns with core): this makes the test run faster and easier to debug in chrome.
- When enabling eager mode, deserialization of mobilenet in the browser fails. Filed https://github.com/tensorflow/tfjs/issues/1060

BUG, PERF

![bench](https://user-images.githubusercontent.com/2294279/50600901-8f0b5c80-0e80-11e9-94f7-22e818cc45e0.png)