Fix firefox glsl 300 failure to compile large constants in clip_gpu (#1480)
The glsl 300 change also caused clip tests in layers to fail. In losses.ts, there is a call to clip(x, min, Number.MAX_VALUE) and Number.MAX_VALUE was spliced into the shader code causing compilation to fail.

The fix was to upload min and max as uniforms, which also reduces shader re-compilation.