Try not to laugh project with Tensorflow.js
Hello, all!

I made Try not to laugh with face-api.js which is based on tensorflow.js.
This project is based on face landmark detection. 
It detects your smile by checking face landmark.
So you have to endure smiling until video end.
I made it for fun and it will be good example for application of tensorflow.js.