Make defaultValue optional in sparseToDense (#1560)
BUG

I'm not sure I have all the context on this op but it looks like defaultValue should be an optional parameter https://www.tensorflow.org/api_docs/python/tf/sparse_to_dense. If it's required then the example in the docs should be updated since it currently throws an error.