Remove OS X testing from travis, only test on browserstack on node 10. (#179)
DEV
This will make travis much faster (we're running 4x as many browser tests as we need to, which will hog browserstack).

OS X testing is not necessary, we test on real browsers on OS X on Browserstack. We only need to test the browser tests from a single build config as well.