webgpu: matmul_packed: fix WPT so it doesn't do unused work (#1731)
FEATURE

When WPT was >1, we were dispatching WPT*WPT times as many threads as we
should have been.

This fixes the strange performance characteristics of WPT, so WPT is now
defaulted to 4. (4 is fastest on my machine, but 8 crashes.)

I did a lot of renames throughout the code to make the naming less
muddy.

The only other functional change here is that, in matmul_webgpu,
localRow and localCol are transposed (y,x instead of x,y) to match
matmul_packed_webgpu. This makes it much faster and brings the
performance results more in line with expectations:

Naive: 5.1ms
Packed, WPT=1: 5.0ms
Packed, WPT=2: 3.1ms
Packed, WPT=4: 2.5ms