Add benchmark for tf.GraphModel (#1580)
This PR adds benchmarks for  the GraphModel version of MobileNetV2.

Such models are generated from Python TensorFlow in the following steps:
1. Load the MobileNetV2 application from tf.keras
2. Save them as SavedModels using `tf.keras.experimental.export_saved_model()`
3. Convert the SavedModels into tfjs_graph_model format using tensorflowjs_converter

The during the browser (karma) benchmarks, the tfjs_graph_model artifacts are loaded in the
testing page and their `predict()` methods are called with random data for performance benchmarking.

Like other model formats (LayersModel), the results are logged to Cloud Firestore, which can
be viewed in the frontend. 

For MobileNetV2, the four width values are benchmarked: 0.25, 0.5, 0.75, 1.0.
For each of the width values, the benchmark numbers consistently show a performance
advantage of GraphModel over LayersModel by about 15%-20% in the browser and 30%-40% in Node.js.