Fix unit tests to ensure that test registrations happen before tests. (#1788)
Previously the test registries would happen as a function of imports. This doesn't ensure that unconstrained tests run before tests get registered.

This PR creates a `setup_test.ts` file like we do in WebGPU which imports the registries and then parses the karma flags. This ensures the expected order.

This PR also adds --testEnv as a command line flag instead of --backend.
DEV