[tfjs-layers] Fix a silent error in L1/L2/L1L2 regularizer constructor
Re. https://github.com/tensorflow/tfjs/issues/1773

BUG

Previosly, if you pass a number argument to `tf.regularizers.l2()`,
for example, no error will be thrown, even though the specified number
will not take effect as the regularization coefficient.

This PR fixes that by explicitly checking the type of the arg.

This is applicable to:
- tf.regularizers.l1()
- tf.regularizers.l2()
- tf.regularizers.l1l2()