Fix memory leak with loaded models containing nested Sequential layers
If a model loaded with loadLayersModel contains a nested Sequential layer,
disposing  of the main model leaves the tensors of the nested model
behind. This is because the refCount of the nested model doesn't go to zero
during the call to `dispose` since the refCount is incremented twice:
once during submodel creation and once when the submodel is added as a layer.

This commit makes it so references to each internally constructed sub-container
are tracked and disposed when the parent model is disposed.