webgpu: Fix argMax benchmark doesn't work
BUG

Fix below error message when you run 'yarn benchmark':
  TypeError: Cannot read property 'data' of undefined

The error is because we want to test 50 sets of data. However, we used the
default 'reps' which is 1, that will result only the first element of
maxes is valid. So we will get above error message for the following
elements in maxes.
Meanwhile, decrease n to 2 to avoid out of time.