[WASM] Add kernel registry (#2172)
FEATURE

Add kernel registry which enables us to register a kernel without having to extend a centralized `Backend` interface. This allows us to **gradually** modularize our backends, where each kernel implementation lives in a separate file and can be easily excluded from the bundle.

As a proof of concept extract the `tf.square` op and its CPU kernel into a separate file. The goal is to test this registry as we build out our WASM backend.

To maintain backwards-compatibility, we extend `engine.runKernel()` to take optionally a kernel name in addition to the function closure. If the kernel is found in the registry, the engine will execute that function, else it will fall back to calling `backend.op()`. This way backends like webgl/node/webgpu continue to work while the wasm backend is modular.

Details:
- Each kernel has an official name defined in the global TF registry - see [`ops.pbtxt`](https://github.com/tensorflow/tensorflow/blob/master/tensorflow/core/ops/ops.pbtxt), which we will follow.
- Add API to register a kernel `tf.registerKernel(kernelName, backendName, kernelFunc);`
- Extract the `Square` kernel and its higher-level op into separate `src/backend/cpu/square.ts` and `src/ops/square.ts` files respectively.
- This change is a no-op since we explicitly import `square.ts` when creating the bundle.