Add a CI for checking the WASM .wasm and .js size (#2191)
DEV

Here we move some of the bundle size checking utilities from core out into the union package so we can share logic with the WASM backend.

Also this PR moves `yarn test-cc` into a separate cloud build job so that it parallelizes with the browser tests.

The bundle size output looks like this:

```
$ ./scripts/test-bundle-size.js
~~~~WASM file ~~~~
==> post-gzip
  diff:      0.00 K  (0.00%)
  master:   59.39 K
  change:   59.39 K

==> pre-gzip
  diff:      0.00 K  (0.00%)
  master:  183.49 K
  change:  183.49 K


~~~~minified bundle (JavaScript) ~~~~
==> post-gzip
  diff:      0.00 K  (0.00%)
  master:    7.58 K
  change:    7.58 K

==> pre-gzip
  diff:      0.00 K  (0.00%)
  master:   20.99 K
  change:   20.99 K
```

Today we are at 66.97K post-gzip.

As of this change the cloudbuild.yml is *disabled* for this because it relies on the new package structure. After this change I'll reenable it and fix any bugs.