[Core] Fix bug with WEBGL_PACK=false and uploading packed tensor (#2291)
BUG

Fix a bug with WEBGL_PACK=false and uploading packed tensor. Also fix a bug in how we count data ids in the backend. This makes all but 1 test to pass with WEBGL_PACK false.

Since I'm out of battery, will address the remaining unit test in another PR where I will enable WEBGL_PACK false for CI testing. The remaining test is NOT critical - it is related to debug mode.