[core, wasm] Modularize `FromPixels` and make it work for the w… (#2429)
Remove `backend.fromPixels()`. The user-facing `tf.browser.fromPixels()` will call the current backend's `FromPixels` kernel IF it is registered. Else it will fallback to a platform-agnostic and backend-agnostic logic in a centralized place.

The `cpu`, `wasm` and `node` backends do not need to have `FromPixels` kernel registered since the centralized logic in `tf.browser.fromPixels()` will work.

The desired outcome of this is that now `tf.browser.fromPixels()` just works with the WASM backend without any additional code.

FEATURE