[WASM] Fix bug where we try to set more memory on WASM heap tha… (#2591)
BUG

If a values TypedArray is passed to the WASM backend we:
1) malloc bytes based on the shape, dtype
2) set the buffer based on the values buffer size.

However, the values buffer size could not match the size of the values array itself because values can be a view of a larger buffer (e.g. with Array.subarray()).

This makes sure we only set into the WASM heap the number of bytes we've actually malloc'd.

The unit test I wrote to catch this is now in core (as all backends need this to be true) and I verified it crashes the page without the fix.