webgpu: Fix the mapAsync error
BUG

This change fixes below error:
Uncaught (in promise) DOMException: Unknown error in mapAsync.

Since f is an async function, we should use 'await f()' to make sure the promise has finished.
Otherwise, the execution sequence of function in and out of the body will mess and generate weird result.