Update FromPixelsImageBitmap Pipeline Cache and Uniform Strategy
In previous code, FromPixelsImageBitmap will cache the whole program
when it has been created first time and keep it unchanged.

But preprocessor will change the shader code based on output shape. So
the cache strategy of FromPixelsImageBitmap is not correct.

In previous code, the uniform buffer will always be destroyed. Since
uniform data length is stick, we can reuse the same buffer always but
update the content.

This CL updated the strategy to use WebGPU pipeline cache system and
check out shape for every invocation. And reuse the uniform buffer after
it is created.