webgpu: Use WEBGPU_COMMAND_ENCODER_COUNT_IN_QUEUE
PERF

This patch uses WEBGPU_COMMAND_ENCODER_COUNT_IN_QUEUE flag instead of
the old WEBGPU_IMMEDIATE_EXECUTION_ENABLED flag. When
WEBGPU_COMMAND_ENCODER_COUNT_IN_QUEUE is one, it will become to the old
immediate mode.

queue.submit has a fixed cpu overhead. So it's not efficient to call it
too many times. It's also not a good balance if we just submit once for
each frame. Here we use an empirical value 15, which is observed over 5%
performance improvement for mobilenet/resnet50 on different platforms.