e2e: Put profileMemoryAndKernelTime after measureAveragePredictTime (#4929)
* e2e: Put profileMemoryAndKernelTime after measureAveragePredictTime

This PR removes the sleep time and puts profileMemoryAndKernelTime after
measureAveragePredictTime so that the sum of profiling kernel time is
close to the average CPU time. This is because that after many iterators
the CPU/GPU frequnecy has reached to the highest. And at that point to
query the kernel time, it is more reflective of the real ops time in the
model.

This PR also adds CHECK_COMPUTATION_FOR_ERRORS flag to get more accurate
profiling result when disabling it.

* Clear the showing message at the end of profileMemoryAndKernelTime

Co-authored-by: Na Li <linazhao@google.com>