Create a separate 'link-package-webgpu' for webgpu
webgpu is often built on Windows, but tfjs-tflite does not currently build on Windows due to requiring wget. tfjs-tflite is included in the 'link-package', which webgpu depends on for tfjs-core and others. This PR creates another link package for webgpu, named 'link-package-webgpu', that links only the dependencies that webgpu needs to build, avoiding the tflite incompatability.