Base the release docker off of debian (#5640)
This image bases off of debian:10 instead of debian:11 because gcc-10 in debian:11 segfaults on BatchMatMul and FusedMatMul cc tests in the wasm backend. Debian:10 uses gcc-8 which does not have this problem.

This PR updates the dockerfile, but it does not update the cloudbuild files to use the release image because the image must be published first.