[tflite] Fix the broken demo (#5732)
The demo was broken because the wasm files (from the tfjs-tflie package) were not copied to the build target directory so the demo cannot load the wasm loader/binary (server would return a 404 html page).

This PR added the missing staticFiles object to package.json in order to specify which source directory to copy files from to the build target directory (dist/). This functionality is supported by the parcel-plugin-static-files-copy npm package which is already there.