[webgl] GatherV2 fill out of range values with zero
GatherV2 used to throw an error when accessing an index out of range, but that requires a copy back to the CPU. This change improves performance and brings the API in line with the [C api](https://www.tensorflow.org/api_docs/cc/class/tensorflow/ops/gather-v2). For debugging purposes, the original behavior can be re-enabled via the DEBUG flag.