Update to wording for starting values of indices
To avoid confusion. User's should be aware that indices must start from 0 when encoding their class data.
User can not use indices that start from say 1.

So if you have 2 classes of data, Class 1, and Class 2:

[1,2,1,2] is not valid.

Indices must be encoded as:

[0,1,0,1]

For class 1 (0) and class 2 (1) respectively for 1 hot to work as expected.

This is not obvious or expected in current documentation.