Downgrade .bazelversion to 4.2.2 to fix file path issue
File paths in tfjs-converter/scripts/kernels_to_ops.bzl assume that the `bazel-out` symlink directory exists at the root of the workspace.

```python
"../" + output_file.path, # output_file.path starts with `bazel-out/`
```

This might be a bug in Bazel, or we're just doing something unexpected here.

Bazel 5.0.0 renames this to ${workspace prefix}-out (dist-out in our case), so the file is not found.