[wasm] Fix threaded worker blob syntax error
JS's Blob replaces "\n" with a newline character. In the threaded worker bundle, there's a "\n" in `if(ENVIRONMENT_IS_NODE){fs.writeSync(2,text+"\n")`. When it is replaced with a newline, it splits a string across multiple lines, which is not valid JavaScript.

This change prevents Blob from doing that replacement by escaping the "\n" character as "\\n".

fixes #6203