[webgpu] Migrate sparseToDense to the atomic-based kernel
PERF

The sparseToDense op takes an optional default value. Unlike scatterNd,
the output cannot be initialized with fill(), since the default value is
a scalar tensor (which could be the result of a previous op) than a
scalar number. The (horrible!) workaround here is to broadcast the value
with tile().

The other challenge is if the kernel should discard the original value
at index or accumulate on that. The magic is performed by splitting the
op into two "scatter" steps: 1) replace the default value with 0 or add
0 to 0, and 2) add the input sparse values to 0 or whatever. This avoids
a bitmap for recording whether the output element at index has been
updated by another invocation.