Add a Bazel test_suite to organize tests (#6550)
Create a test_suite target named tests in each package and a test_suite in the root of the repository that runs all the tests for all the Bazel packages. WebGPU is excluded at the moment because it is in alpha.

Running yarn test at the root of the repository will run the tests.

This target is just for developer convenience.